+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@
 
 * SubSonic - When calling createPlaylist, assume that the list needs to be empty first
 * Simplify flagging/rating multi-disk albums
+* SubSonic - just send getmusicfolders music folders
 
 ### Fixed